backend/requirements.txt
alembic==1.17.2
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
Flask-CORS==4.0.0
flasgger==0.9.7.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mistune==3.1.4
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
pytest==8.3.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.4
